requirements.txt
gradio==4.44.1
telethon==1.36.0
python-dotenv==1.0.1
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
pgvector==0.2.5
pydantic==2.8.2
tenacity==8.5.0
sentence-transformers==3.0.1
numpy==1.26.4
pandas==2.2.2
openai==1.45.0
